No sleeping after write operations (#16)
* No sleeping after write operations

After a write operation, if the work rate is sufficiently high then there is a 25ms sleep.  It appears that the purpose of this is to give some time for the sync interval to elapse and flush the pending updates, and maybe to slow down spikes of writes.

The problem with this is that work is not getting done during sleep unless the sync interval elapses.  Instead of sleeping it would be more productive to send a signal to the run goroutine to tell it to do a flush immediately. Also, if the sync interval is very long, then writers may get very slow as they will do a lot of sleeping.

This PR replaces possible sleeping after a write operation with possibly sending a signal to do an immediate flush, without waiting for the signal to be read.  Slowing spikes in write activity until data is flushes now relies on synchronization between the writes and flush, instead of time delay, to generate backpressure.

This PR also includes some minor changes to reduce locking.