Compact by removing from front of index
There is very little benefit to removing entries from the index other than from the front.  This greatly simplifies tracking dead entries to remove and the process of compaction.  A single checkpoint is advanced from the front of the index until it encounters an entry still in use.  When the index has advanced far enough into the index, compaction is performed.