Store the index across multiple file and implement GC
The index is stored across any number of files each 1GB is size.  The index is an appending log and appends to the latest index file until that file has reached its size limit (1GB), at which point a new index file is created.  Only the newest index file is written to.

Periodically, garbage collection (GC) is run to clean up files that only contain stale (no longer in use) entries.  The garbage collector deletes the stale files.  This keeps the index data to only what is active, without having to copy any index data to a compacted file.