Fix corrupted index files (#30)
* Fix corrupted index files

It is possible for an index file to have incomplete data if a process is terminated unexpectedly.  This will cause an error reading the bad index file and it will not be fixed until file is repaired or deleted.  This change removes the damaged portion of the index file and allows the remaining data in the file to continue being used.

* bump version