Flush write buffer when Flush is called
When Flush was called on the index or the primary, the write buffers were not flushed, so the index or the primary may be written to disk and the other not.  With a hard shutdown this could leave the database in an inconsistent state that can cause errors.  Now the write buffers are flushed when calling Flush.  This helps the situation significantly, but is not a complete solution.  For that, writes to the index and primary would need to be done as a transaction.

Other changes:
- Removed unnecessary exports
- Comments
- Modified buffer size