Flush write buffer when Flush is called (#37)
### Flush write buffer when Flush is called

When Flush was called on the index or the primary, the write buffers were not flushed, so the index or the primary may be written to disk and the other not.  With a hard shutdown this could leave the database in an inconsistent state that can cause errors.  Now the write buffers are flushed when calling Flush.  This helps the situation significantly, but is not a complete solution.  For that, writes to the index and primary would need to be done as a transaction.

### Repair corruption from lost keys

If a bad key is read from the primary, this means that either that data in the primary is bad or the index points to an invalid location in the primary.  Either way, the index record is not usable, so replace that index record with a new record that points to a new location in the primary.

Other changes:
* Removed unnecessary exports
* Comments
* Modified buffer size
* Flush freelist on close. Add index diagnostics.
* Test bad index data
* Do not create GC channels if GC is disabled