Fix a data race and data loss error in flush
There are two pools, a read (curPool) and and write (nextPool). When Flush is called these pools are swapped and the curPool is written to disk.

Race: If there are concurrent calls to Flush, the 2nd Flush can swap the pools while the 1st Flush is still reading the pool being flushed. That could
cause the pool being read by the 1st Flush to be written to concurrently. This is fixed by only allowing one flush at a time.

Data Loss: After Flush swaps the pools and writes the curPool, it does not reset that pool. This causes subsequent Flushes to unnecessarily write data to disk, and is may also cause data loss if the there is newer data in the other pool. This is fixed by resetting the flushed pool before flush returns.