Fix a data race and data loss error in flush (#55)
* Fix a data race and data loss error in flush

There are two pools, one to flush (`curPool`) and one to update (`nextPool`). When `Flush` is called these pools are swapped and the items in `curPool` are written to disk.  This pertains to the index, primary, and freelist.

### Race:
If there are concurrent calls to `Flush`, the 2nd `Flush` can swap the pools while the 1st `Flush` is still reading the pool being flushed. That could cause the pool being read by the 1st `Flush` to be written to concurrently. This is fixed by only allowing one `Flush` at a time.

A second race was present between concurrent `Flush` and `Sync`, since `Sync` could reset the `curPool` while `Flush` was still reading it. This is fixed by resetting the `curPool` only in `Flush`.

### Data Loss:
`Flush` swapped the pools and flushed `curPool`, and did not reset either pool. Having already flushed data in the update pool caused subsequent calls to `Flush` to unnecessarily write that data to disk again. Swapping pools and retaining old data may also cause data loss if there is a newer version of the same data in the other pool; flushing the pool with the older data may replace the newer data. This is fixed by not swapping the pools, and instead creating a new pool for new updates to be written to

NOTE: Both of these issues were present for all history of storethehash. These problems were mostly avoided by automatically `Flush` at the store level explicitly or automatically, which callis `Sync` and resets `curPool`. Concurrent calls to `Flush` could still have caused problems.

* Fix index
* Fix multihash primary
* Fix cid primary
* Fix freelist
* Add test for index flush race
* Add test for multihash primary flush race
* Add test for cid primary flush race
* Add tests to check that Flush does not write excess data