Define store options in store package
The options defined for the blockstore all pertain to the underlying store, which is where the options really need to be defined. This allows users of the store package to be able to use to options as well.

Add a GCTimeLimit option to limit the time GC is allowed to run.