More aggressive GC for index files (#59)
* Aggressive index garbage collection

- Remove all files not referenced by buckets
- Truncate free spans at end of index files
- Merge free spans withing index files

* Test truncate and merge
* Log records marked as free and merged
* Save in-memory buckets state on Close

Instead of scanning index files, load the saved bucket in-memory state. This makes reloading nearly instant. If the saved bucket state is not available, fall back to scanning index files. 

* Define store options in `store` package

The options defined for the blockstore all pertain to the underlying store, which is where the options really need to be defined. This allows users of the store package to be able to use to options as well.

Add a GCTimeLimit option to limit the time GC is allowed to run.

* Comments on options
* Do not log GC timeout as error
* Change repetitive names