Do not uses in-memory size buckets
The size buckets were being used in exactly one place, that it, when reasing the record list from an index file. The size of the record is already present at the place where that data is read, so read it from there instead of keeping the size in memory.

Removing the size buckets saves considerable memory (2 ^ bitPrefxSize) * 4 bytes at no cost to performance.