Use LRU file cache for reading index files (#72)
* Use LRU file cache for reading index files

When reading values for multihashes, it is necessary to first open the index file that has the index record for the multihash's value list. When accessing many multihashes, this results in a large number of syscalls for opening and closing the index files. Statistically, more of the indexes reside in more recently created index files, so on average only a subset of the total files is heavily accessed. By using an LRU cache to keep some number of these files open, many of the syscalls can be avoided.

* If the cache size is set to zero, then this reverts the behavior to opening and closing files directly.
* Keep track of hits and misses, and print stats on each GC cycle.
* API to set file cache size at run-time