Wait for flushes to catch up with incoming data
If the rate of incoming work exceeds the rate that work can be flushed at, and there is enough work to be concerned about, then trigger an immediate flush and wait for the flush to complete. It is necessary to wait for the flush, otherwise more work could continue to come in and be stored in memory faster that flushes could handle it, and this could lead to memory exhaustion.