Fix race condition when calling Sync
If is possible that when Sync is called on a file, that the file has been closed and the new file has not yet been opened. This will cause sync to return an error, which in turn will cause all other operations on the file to fail. This race affects the index, the primary, and the freelist. A mutex is used to fix this by making the closing of the old file and the opening of the new file atomic from the view of Sync.

Make calling Sync as part of Flush configurable. For most applications, it probably does not improve durability to call fsync, unless there are certain guarantees about the state of storage that are being made at certain points in time. Performance can be improved by not calling fsync if it is not necessary.