Fix an error when closing a file in the file cache (#80)
* Fix an error when closing a file in the file cache

A "file already closed" was happening in the following situation:

Open a file to get 1st *File, let it be evicted, open that same file again to
get 2nd *File, then close both 1st and 2nd *File. The first Close was closing the one in the cache, when it should have been closing the one in the evicted but still open file. The second close tries to close the file in the cache and, results in a "file already closed" error.

This PR ensures that the first close closes the evicted *File, not the one in cache.