Fix unindexed primary records (#84)
* Fix unindexed primary records

When GC tries to move a primary record from a low-use primary file, it needs to update the index of the record to the new location. If the index for the record being moved does not exist, it means there the index was deleted but the record was not marked as deleted, or a record was added to the primary and the index was never updated. It is impossible to tell what type of corruption caused the situation so it is unclear whether to add the index back, or remove the unreachable record.

The choice here is to err on the side of removing data. It is better to have some missing data that can eventually be replaced, than have unreachable data that can never be removed. This PR deletes the primary record if it finds its index is missing.

This fix is necessary to allow primary GC to continue when this form of corruption is encountered.

* freelist processing needs to check the context